requirements.txt
asgiref==3.2.10
Django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2020.5
whitenoise==5.2.0
django-filter==2.4.0
Markdown==3.3.3
djangorestframework-gis==0.16
